back/requirements.txt
albumentations==1.4.4
flask==3.0.3
gunicorn==22.0.0
matplotlib==3.8.4
numpy==1.26.4
opencv-python==4.9.0.80
pandas==2.2.2
pillow==10.3.0 
scikit-image==0.22.0
scikit-learn==1.4.2
Werkzeug==3.0.2
scipy==1.13.0